Fix tests

+++ hardware_interface/include/hardware_interface/handle.hpp
@@ -104,7 +104,7 @@ public:
       {
         value_ptr_ = nullptr;
         value_ = initial_value.empty()
-                   ? std::numeric_limits<double>::quiet_NaN()
+                   ? std::numeric_limits<float>::quiet_NaN()
                    : static_cast<float>(hardware_interface::stod(initial_value));
         float val = std::get<float>(value_);
         if (val < std::numeric_limits<float>::lowest() || val > std::numeric_limits<float>::max())
